requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.28.1
aiosqlite==0.20.0
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0